Transformations tool (WIP)
Not finished yet, aka there's no link between actual transformations datas in character on screen, and the datas shown in the tool.
But UI seems to work, I can pick-up the changes in the sliders, and it seems that importing data from model or a file is not complicated, most of methods are already written (but not used yet).
So the tool is far from its finish. But you can already play with GUI, to see what's going on.