MB-Lab: Import expression shape keys update
Addressed two comment:
1. removed indent=2 for json.dump
2. renamed the tool to Import instead of Export

I also added an option to override existing shapekeys.
This is important since it allows the user to override their own
shapekeys. It can be expanded later to expose a list of all the
expression shape keys, which the user can modify. However,
I still don't think it's a good idea to override the default
shape keys, as they are pretty good already.

Signed-off-by: Amir Shehata <amir.pavlo@gmail.com>