Fixed a bug for separated projects
With the new "Project" separation, using the creation tools fails, as the config and blend where handled differently. Now it's good, the creator does nothing for the config, and blend must be in its dedicated folder with the good name.
By the way, this name may change in the future, I think that it's better that it has the name of the project like toon_library.blend, not just humanoid_library.blend.